+++ examples/resources/pingdirectory_password_generator/import.sh
@@ -1,0 +1,2 @@
+# "passwordGeneratorId" should be the id of the Password Generator to be imported
+terraform import pingdirectory_password_generator.myPasswordGenerator passwordGeneratorId